convert to i64 before subtracting in musiplayer.is_finished()

+++ musiplayer/src/lib.rs
@@ -110,7 +110,9 @@ impl Player {
         // it does not have a .is_finished() method (or atleast i could'nt find it)
 
         if self.paused || self.duration == 0 {return Ok(false)}
-        Ok(self.duration()?-self.position()? < 50)
+
+        // i64 was needed as in release mode there are no overflow checks and u64-lil_bigger_u64 cant be smaller than 50
+        Ok((self.duration()? as i64) - (self.position()? as i64) < 50)
     }
 
     fn is_paused(&self) -> bool {
